src/consumer/requirements.txt
paho-mqtt==2.1.0
pydantic==2.12.0
pymongo==4.10.1
slack-sdk==3.37.0
dotenv==0.9.9

